Proyecto/backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-HTTPAuth
python-dotenv
Flask-Cors
Cors
marshmallow
marshmallow-sqlalchemy